Library/Mobile Documents/com~apple~CloudDocs/Desktop/scratch/python/EmployeePortal/requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.5.0
appdirs==1.4.4
asgiref==3.5.2
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.6.2
boto3==1.27.0
botocore==1.30.0
bs4==0.0.1
CacheControl==0.13.1
cachetools==5.3.1
catalogue==2.0.9
certifi==2022.5.18.1
cffi==1.16.0
charset-normalizer==2.0.12
ChatterBot==1.0.4
chatterbot-corpus==1.2.0
click==8.1.7
confection==0.1.1
constantly==15.1.0
crispy-bootstrap4==2022.1
cryptography==41.0.1
cssselect==1.1.0
cymem==2.0.7
daphne==4.0.0
dateparser==1.1.1
defusedxml==0.7.1
demjson3==3.0.5
Deprecated==1.2.14
distlib==0.3.6
Django==4.1.2
django-allauth==0.54.0
django-bootstrap4==23.2
django-crispy-forms==2.0
django-filter==23.2
django-oauth-toolkit==2.3.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
django-simple-history==3.3.0
django-storages==1.13.2
django-utils-six==2.0
fake-useragent==0.1.11
filelock==3.9.0
frozenlist==1.3.3
grpcio==1.56.0
grpcio-status==1.56.0
httplib2==0.22.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.11.4
incremental==22.10.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
langcodes==3.3.0
lxml==4.9.0
MarkupSafe==2.1.3
mathparse==0.1.2
msgpack==1.0.5
multidict==6.0.4
murmurhash==1.0.9
nltk==3.8.1
oauthlib==3.2.2
packaging==23.1
parse==1.19.0
pathy==0.10.2
Pillow==9.2.0
Pint==0.22
platformdirs==2.6.2
preshed==3.0.8
proto-plus==1.22.3
protobuf==4.23.3
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.22
pydantic==2.2.1
pydantic_core==2.6.1
pyee==8.2.2
PyJWT==2.7.0
pymongo==3.13.0
pyOpenSSL==23.2.0
pyparsing==3.1.0
pyquery==1.4.3
python-dateutil==2.7.5
python-decouple==3.8
python-dotenv==1.0.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==3.13
regex==2022.3.2
requests==2.27.1
requests-html==0.10.0
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.1
service-identity==23.1.0
six==1.16.0
smart-open==6.3.0
soupsieve==2.3.2.post1
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
sqlparse==0.4.2
srsly==2.4.7
thinc==8.1.12
tqdm==4.64.0
Twisted==22.10.0
txaio==23.1.1
typer==0.9.0
typing_extensions==4.7.1
tzdata==2022.1
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.6
virtualenv==20.17.1
w3lib==1.22.0
wasabi==1.1.2
whitenoise==6.5.0
wrapt==1.15.0
yarl==1.9.2
zipp==3.8.0
zope.interface==6.0